+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changelog
 ------------------
 
 - #2848 Patch transaction commit hook for catalog queue to avoid double indexing
+- #2844 Allow unit selection in profiles
 - #2846 Fix Unit super-/subscript rendering 
 - #2847 Remove nonexisting getSamplePointTitle column index
 - #2776 Refactor: rename SiteView.setCookie to set_cookie in dashboard.pt